VMI: Split hotplug logic to its own file
